fix(rules): validator/API consistency batch (placeholders, Doppler, boundaries, Stats types)
Batch of small verified inconsistencies from the pre-publish review:

1. SlackUserToken skipped the placeholder checks on its digit segments
   that SlackBotToken already performs, so a byte-identical placeholder
   shape confirmed under one trigger and was rejected under the other.
   Now checked identically.
2. DopplerToken applied the dp.st.-only optional config-label grammar to
   all seven scoped prefixes, over-matching bodies like "dp.pt.dev.<40
   chars>" that Doppler's own format regex doesn't allow for that prefix.
   Now gated on the matched trigger being exactly "dp.st.".
3. Consolidated on one placeholder predicate (isPlaceholder, sourcecontrol.go)
   across every validator that previously called entropy.IsPlaceholder
   directly, removing the drift risk between two behaviorally different
   predicates for the same job. Fixed a stale file-header comment in
   ai.go claiming isPlaceholder lives in validators.go.
4. Applied one boundary convention (maxLookbehind: 1 + a preceding-
   identifier-byte rejection) across every fixed-literal-prefix token
   rule that lacked it (GitHub PAT/OAuth/App/refresh/fine-grained,
   GitLab PAT/runner/deploy/feed/pipeline-trigger, npm, PyPI, Docker Hub,
   Stripe secret/webhook/ephemeral, Vault service/batch, Vercel,
   Doppler), matching the convention AWS/GCP/Hugging Face/Groq/OpenAI/
   Cursor rules already used. Documented the convention in
   docs/RULE_AUTHORING.md.
5. Unified Stats.Findings (was int) and Stats.ByRule's values (was
   map[string]int) onto int64, matching every other Stats counter. This
   is a public API change landing before the first public tag as noted
   in the review.
6. Documented that Config.OnFinding is invoked concurrently across
   concurrent Redact calls sharing one Engine, and that Redact's ctx is
   polled once per read iteration (cannot interrupt a blocked Read), on
   both the internal and public API doc comments.

No accuracy-corpus false-positive regressions; fixtures added for items
1, 2, and 4's reproduction cases.

Resolves ENG-190

+++ internal/goredact/engine.go
@@ -406,7 +406,7 @@ func (r *scanRun) recordFindings(released []span.Span) {
 		r.stats.Findings++
 		id := r.e.rules.Rules[s.Rule].ID
 		if r.stats.ByRule == nil {
-			r.stats.ByRule = make(map[string]int)
+			r.stats.ByRule = make(map[string]int64)
 		}
 		r.stats.ByRule[id]++
 		if r.e.cfg.OnFinding != nil {
